refactor(authz): re-organize files
