+++ packages/nestjs-authorization/src/conditions/condition.service.ts
