+++ packages/nestjs-authorization/src/permissions/check-permission.decorator.ts
@@ -1,5 +1,5 @@
 import { applyDecorators, UseGuards } from '@nestjs/common';
-import { Action } from './action';
+import { Action } from '../actions/action';
 import { PermissionGuard } from './permission.guard';
 
 export function CheckPermission(...actions: Action[]) {
