+++ packages/nestjs-authorization/src/permissions/noop-permission-provider.ts
