+++ packages/nestjs-authorization/src/permissions/permission.provider.ts
@@ -1,7 +1,5 @@
 import { PermissionSet } from './permission-set';
 
-export const PERMISSION_PROVIDER = 'PERMISSION_PROVIDER';
-
 export interface PermissionProvider {
   getPermissionSet(ctx: any): Promise<PermissionSet>;
 }
